Case SystemCoreClock to unsigned int
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@4248 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ demos/test_conio.c
@@ -22,7 +22,8 @@ int main(void)
 
   printf("\033[H\033[2J%s Console I/O Test (" __DATE__ " " __TIME__ ")\n\n", MCUFAMILYNAME);
   puts(revision);
-  printf("\nCPU Freq:%d Hz  Compiler:%s %s %s\n\n", SystemCoreClock, __COMPILER__, __VERSION__, __ABI__);
+  printf("\nCPU Freq:%u Hz  Compiler:%s %s %s\n\n", (unsigned int) SystemCoreClock,
+    __COMPILER__, __VERSION__, __ABI__);
 
 // Test putchar()
